refactor(persistence): break audit<->persistence_db cycle, move events SQL
Extract pure event/envelope/audit-log types into a new leaf crate
audit_types (no SQL, no runtime). audit now depends on audit_types AND
persistence_db; persistence_db depends only on audit_types — acyclic
(audit -> persistence_db -> audit_types <- audit).

Move the events-table SQL out of audit::bus::EventBus into
persistence_db::repositories::events (free fns over &SqlitePool). EventBus
delegates publish/replay to those fns and implements the new
audit_types::EventPublisher trait so persistence_db::lifecycle can publish
without depending on audit's EventBus concretely (SqliteLifecycleRepository
now takes impl EventPublisher, stored as Arc<dyn EventPublisher>, so the
public type stays non-generic and existing bare-type call sites elsewhere
in the workspace are untouched). Publish still happens after tx.commit(),
unchanged atomicity.

Scaffold 8 empty q_* repository stub modules (one per db-boundary-zero
drain node) plus make persistence_db a production dependency of
targeting_resolver, so drain nodes never collide on shared files.

crates/audit/src/bus.rs now has 0 raw-sqlx sites (was 7).
