+++ crates/audit-types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "audit"
+name = "audit_types"
 version = "0.1.0"
 edition.workspace = true
 license.workspace = true
@@ -7,19 +7,17 @@ license.workspace = true
 [lib]
 path = "src/lib.rs"
 
+# Pure event/envelope/audit-log type definitions, split out of `audit` so
+# `persistence_db` can depend on the types without depending on the `audit`
+# crate's EventBus runtime (which itself depends on `persistence_db` for its
+# SQL). See docs/development/persistence-layer-hardening.md.
 [dependencies]
 async-trait.workspace = true
-audit_types = { path = "../audit-types" }
 domain_core = { path = "../domain/core" }
-persistence_db = { path = "../persistence/db" }
 schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 specta.workspace = true
-sqlx.workspace = true
-thiserror.workspace = true
-time.workspace = true
-tokio = { workspace = true, features = ["sync", "macros", "rt"] }
 
 [lints]
 workspace = true
