+++ crates/persistence/db/src/repositories/q_projects.rs
@@ -1,0 +1,3 @@
+//! Scaffolding stub for the db-boundary-zero drain node owning `q_projects`.
+//! Populated with free-fn repository queries as that node's raw-sqlx sites
+//! are migrated out of production code.
